Refactor registration bot
